move hidden blueprint from yafowil.common to yafowil.hidden
